fix(events): fix keyboard/mouse events issues when running in another window
When a note is open in a separate window, instanceof checks fail for types like MouseEvent,
KeyboardEvent, etc. With this fix, they work as explected.

- In keyboard-shortcuts.ts although addEventListener requires a callback with a param of type
Event, it is the derived UIEvent that defines the .instanceOf method. I'm not sure if changing
the param type to UIEvent is the proper fix, but it prevents compile errors.

- In droppable.ts there are two instanceof HTMLElement checks, but surprisingly they work as
expected in other windows without any changes. Also the used EventTarget type has no .instanceOf
method, so I left them be.

- There are a few uses of the window and document globals throughout the codebase, which may or
my not need to be replaced by activeWindow and activeDocument - I haven't tested any of them and
I don't' know if they work as expected in other windows

The fix is made to comply with the referenced ObsidianMD blog post.

REF: [How to update your plugin to support pop-out windows]
(https://obsidian.md/blog/how-to-update-plugins-to-support-pop-out-windows/)